keras_conv_net_basics/requirements.txt
tensorflow-gpu==2.0.0-beta1
imutils
matplotlib
numpy
tqdm